Incremental checking of a proof
